test(matrix): refactor tests to use shared InlineExecutorLoop helper
- replace local InlineExecutorLoop with shared helper from tests.helpers
- update imports to import InlineExecutorLoop and matrix_utils functions (send_reply_to_meshtastic, handle_matrix_reply, etc)
- remove obsolete Meshtastic reply tests in test_matrix_utils.py
- extend matrix reply tests with success and numeric-id cases
